+++ Solution/iOS/RocketApp/Features/Sources/Login/AccessibilityKeys+Login.swift
@@ -1,0 +1,7 @@
+import UIToolkit
+
+extension AccessibilityKeys {
+  public enum Login {
+    public static let title = "titleID"
+  }
+}
