--- Solution/iOS/RocketApp/Features/Sources/RocketList/AccessibilityKeys+RocketList.swift
@@ -1,7 +1,0 @@
-import UIToolkit
-
-extension AccessibilityKeys {
-  public enum RocketList {
-    public static let list = "rocketListID"
-  }
-}
