+++ Solution/iOS/RocketAppUITests/Tests/Test.swift
@@ -2,7 +2,7 @@ import XCTest
 
 final class Test: BaseTestCase {
   func testTest() {
-    RocketListScreen(app: app)
-      .checkRocketList()
+    LoginScreen(app: app)
+      .checkLoginTitle()
   }
 }
